// src/interface/gas.interface.ts
export interface gasolinera {
    uid: string;
    id: string;
    nombre: string;
    ubicaciontxt : string;
    ubicacionmaps:string;
    foto: string;
    gasolinaver: string;
    gasolinaroj: string;
    diesel: string;
}

export interface promocion {
    uid: string;
    titulo: string;
    subtitulo: string;
    descripcion: string;
    img: string;
}